+++ relm4-macros/tests/ui/compile-fail/component-missing-widgets.rs
@@ -4,7 +4,7 @@ struct TestComponent;
 
 #[relm4_macros::component]
 impl SimpleComponent for TestComponent {
-    type InitParams = ();
+    type Init = ();
     type Input = ();
     type Output = ();